Add a code of conduct
Using the contributor covenant template, which is a good place to start, 
if not totally fine on its own. Can adjust further as needed.

Using this naming, `.github/CODE_OF_CONDUCT.md`, GitHub will [recognize and display alerts](https://help.github.com/articles/adding-a-code-of-conduct-to-your-project/) to contributors.

Open questions: is there a contact email we can use to report bad behavior?